Observation taxonomy table

This a 975-sample subset of the EMP (Earth Microbiome Project) taxonomy table containing 150bp-length sequences. The samples were picked randomly and evenly across 17 habitats and then evenly across studies in each sample type.

observation_metadata_150bp

Format

Data frame with columns

sequence

Observation-sequence found.

taxonomy1

First taxonomic rank.

taxonomy2

Second taxonomic rank.

taxonomy3

Third taxonomic rank.

taxonomy4

Fourth taxonomic rank.

taxonomy5

Fifth taxonomic rank.

taxonomy6

Sixth taxonomic rank.

taxonomy7

Seventh taxonomic rank.

Source

https://earthmicrobiome.org/

Examples

  head(observation_metadata_150bp)

#>                                                                                                                                                 sequence
#> 1 TACGGAGGGTGCAAGCGTTGTCCGGAATCACTGGGTGTAAAGGGTGTGTAGGTGGATCTGTAAGTCAGAGGTGAAAGCCCACAGCTTAACTGTGGAACTGCCTTTGATACTGCTGATCTTGAGTACTGGAGGGGCTGCTAGAATTCGTGG
#> 2 TACGAAGGGTGCGAGCGTTATTCGGATTTACTGGGCGTAAAGGGCGCGTAGGCGGTTTCCTAAGTCACTTGTTAAAGACCACTTAAACCGGTGGGGATGCGAGTGAAACTGGGAGGCTAGAGGGTAGGAGAGAGAAGTGGAATTCTTGGA
#> 3 TACGGAGGGTGCAAGCGTTATCCGGATTCACTGGGTTTAAAGGGTGCGTAGGTGGGTCTGTAAGTCCGTGGTGAAATCTCCGAGCTTAACTCGGAAACTGCCATGGATACTATAGGTCTTGAATACTGTGGAGGTCAGCGGAATATGTCA
#> 4 TACGGAGGGTGCAAGCGTTATCCGGAATCATTGGGTTTAAAGGGTGCGTAGGCGGTCTAATAAGTCAGAGGTGAAATCCCTACCGCTTAACGGAGGAATTGCCTTTGAAACTATCGGACTTGAAATAGTGTGACGTTGGCGGAATGAGTA
#> 5 TACGTAGGGGGCGAGCGTTGTCCGAAGTTACTGGGCGTAAAGCGCGCGTAGGCGGTCTTCCAAGTCTGGGGTGAAAGGTTCAACGCTCAACGTGAACAGTGCCTTGGATACTGGGAGACTTGAGTCTGGGAGGGGGAAGGGGAATTCCTG
#> 6 TACGAAGGGGGCTAGCGTTGCTCGGAATCACTGGGCGTAAAGCGCACGTAGGCGGACTCTTAAGTCGGGGGTGAAATCCCGGAGCTCAACTCCGGAACTGCCTTCGATACTGAGAGTCTTCGAGTCCGGGAGAGGTGAGTGGAACTGCGA
#>     taxonomy1          taxonomy2              taxonomy3           taxonomy4
#> 1 k__Bacteria   p__Bacteroidetes       c__[Rhodothermi] o__[Rhodothermales]
#> 2 k__Bacteria             p__TM6              c__SBRH58                 o__
#> 3 k__Bacteria   p__Bacteroidetes       c__[Saprospirae]  o__[Saprospirales]
#> 4 k__Bacteria   p__Bacteroidetes                                           
#> 5 k__Bacteria p__Armatimonadetes       c__Armatimonadia  o__Armatimonadales
#> 6 k__Bacteria  p__Proteobacteria c__Alphaproteobacteria      o__Rhizobiales
#>              taxonomy5      taxonomy6 taxonomy7
#> 1   f__Rhodothermaceae            g__       s__
#> 2                  f__            g__       s__
#> 3  f__Chitinophagaceae            g__       s__
#> 4                                              
#> 5  f__Armatimonadaceae            g__       s__
#> 6 f__Hyphomicrobiaceae g__Rhodoplanes       s__