Sample metadata table

This is a meta-data/mapping file containing related information about samples originated from the Earth Microbiome Project.

sample_metadata

Format

Data frame with columns

#SampleID

unique sample identifier (leading hashtag is required by Qiime)

BarcodeSequence

barcode sequence

LinkerPrimerSequence

linker primer sequence

Description

sample description

host_subject_id

unique host or subject (can have multiple samples per host_subject_id)

study_id

study identifier

title

study title

principal_investigator

PI of the study

doi

Digital Object Identifier (DOI) or PubMed ID of primary publication

ebi_accession

European Bioinformatics Institute (EBI) European Nucleotide Archive (ENA) accession number if submitted

target_gene

name of gene amplified (e.g., 16S rRNA)

target_subfragment

name of subfragment of gene amplified (e.g., V4)

pcr_primers

amplicon primer sequences used

illumina_technology

model of Illumina sequencer

extraction_center

where the sample was physically extracted

run_center

where the sample was physically sequenced (CCME=U Colorado Boulder, ANL=Argonne Natl Lab, UCSDMI=UC San Diego, CGS-GL=Wash U, UCD=U Colorado Denver)

run_date

date the sample was physically sequenced

read_length_bp

median read length in bp across study after quality filtering

sequences_split_libraries

number of sequences after demultiplexing with split_libraries_fastq.py

observations_closed_ref_greengenes

number of observations in closed-reference Greengenes table

observations_closed_ref_silva

number of observations in closed-reference Silva table

observations_open_ref_greengenes

number of observations in open-reference Greengenes table

observations_deblur_90bp

number of observations in 90-bp Deblur table

observations_deblur_100bp

number of observations in 100-bp Deblur table

observations_deblur_150bp

number of observations in 150-bp Deblur table

emp_release1

samples with >=1 sequences (split libraries) per sample

qc_filtered

samples with >=1000 observations in CR-GG & CR-Silva & OR-GG & Deblur-90 but excluding controls (all subsets are in this set)

subset_10k

10000 samples with >=10000 CR-GG & CR-Silva & OR-GG and >=5000 Deblur-90, randomly selected and evenly distributed across empo_3 categories and then across studies

subset_5k

5000 samples with >=10000 CR-GG & CR-Silva & OR-GG and >=5000 Deblur-90, randomly selected and evenly distributed across empo_3 categories and then across studies

subset_2k

2000 samples with >=10000 CR-GG & CR-Silva & OR-GG and >=5000 Deblur-90, randomly selected and evenly distributed across empo_3 categories and then across studies

sample_taxid

sample NCBI taxonomy ID

sample_scientific_name

sample NCBI scientific name looked up from taxonomy ID

host_taxid

host NCBI taxonomy ID

host_common_name_provided

host common name provided in the original mapping file

host_common_name

host NCBI common name looked up from taxonomy ID

host_scientific_name

host NCBI scientific name looked up from taxonomy ID

host_superkingdom

host superkingdom looked up from taxonomy ID

host_kingdom

host kingdom looked up from taxonomy ID

host_phylum

host phylum looked up from taxonomy ID

host_class

host class looked up from taxonomy ID

host_order

host order looked up from taxonomy ID

host_family

host family looked up from taxonomy ID

host_genus

host genus looked up from taxonomy ID

host_species

host species looked up from taxonomy ID

collection_timestamp

date and time when sample was collected

country

country where sample was collected

latitude_deg

latitude in degrees

longitude_deg

longitude in degrees

depth_m

depth in meters of sample below surface (earth surface if soil, sea/lake bottom if sediment, lake surface if lake, sea level if marine)

altitude_m

height above surface, usually zero unless the mouse is levitating (either depth or altitude can have a non-zero value but not both)

elevation_m

height above sea level in meters (from georeferencing tool)

env_biome

ENVO biome

env_feature

ENVO feature

env_material

ENVO material

envo_biome_0

level 0 ENVO biome looked up from given ENVO biome

envo_biome_1

level 1 ENVO biome looked up from given ENVO biome

envo_biome_2

level 2 ENVO biome looked up from given ENVO biome

envo_biome_3

level 3 ENVO biome looked up from given ENVO biome

envo_biome_4

level 4 ENVO biome looked up from given ENVO biome

envo_biome_5

level 5 ENVO biome looked up from given ENVO biome

empo_0

level 0 EMPO habitat from EMPO ontology

empo_1

level 1 EMPO habitat from EMPO ontology

empo_2

level 2 EMPO habitat from EMPO ontology

empo_3

level 3 EMPO habitat from EMPO ontology

adiv_observed_otus

observed tag sequences of 90-bp Deblur table rarefied to 5000 sequences per sample

adiv_chao1

Chao1 index of 90-bp Deblur table rarefied to 5000 sequences per sample

adiv_shannon

Shannon index of 90-bp Deblur table rarefied to 5000 sequences per sample

adiv_faith_pd

Faith's phylogenetic diversity of 90-bp Deblur table rarefied to 5000 sequences per sample

temperature_deg_c

ambient temperature in degrees Celcius

ph

pH value

salinity_psu

salinity in practical salinity units (PSU)

oxygen_mg_per_l

oxygen concentration in mg/L

phosphate_umol_per_l

phosphate concentration in µmol/L

ammonium_umol_per_l

ammonium concentration in µmol/L

nitrate_umol_per_l

nitrate concentration in µmol/L

sulfate_umol_per_l

sulfate concentration in µmol/L

Source

https://earthmicrobiome.org/

Examples

  head(sample_metadata)

#>   #SampleID BarcodeSequence LinkerPrimerSequence
#> 1    662.M1    CGAGTCACGATT  GTGCCAGCMGCCGCGGTAA
#> 2    662.M2    GCCATAGTGTGT  GTGCCAGCMGCCGCGGTAA
#> 3    662.M3    GTAGACATGTGT  GTGCCAGCMGCCGCGGTAA
#> 4    662.M4    TGTGGCTCGTGT  GTGCCAGCMGCCGCGGTAA
#> 5   662.M5a    TCTGATCGAGGT  GTGCCAGCMGCCGCGGTAA
#> 6   662.M5b    AGAGAGACAGGT  GTGCCAGCMGCCGCGGTAA
#>                             Description host_subject_id study_id
#> 1  biofilm brushed from crucible lid M1              M1      662
#> 2  biofilm brushed from crucible lid M2              M2      662
#> 3  biofilm brushed from crucible lid M3              M3      662
#> 4  biofilm brushed from crucible lid M4              M4      662
#> 5 biofilm brushed from crucible lid M5a             M5a      662
#> 6 biofilm brushed from crucible lid M5b             M5b      662
#>                                                                            title
#> 1 The role of macrobiota in structuring microbial communities along rocky shores
#> 2 The role of macrobiota in structuring microbial communities along rocky shores
#> 3 The role of macrobiota in structuring microbial communities along rocky shores
#> 4 The role of macrobiota in structuring microbial communities along rocky shores
#> 5 The role of macrobiota in structuring microbial communities along rocky shores
#> 6 The role of macrobiota in structuring microbial communities along rocky shores
#>   principal_investigator               doi ebi_accession target_gene
#> 1      Catherine Pfister 10.7717/peerj.631     ERP020507    16S rRNA
#> 2      Catherine Pfister 10.7717/peerj.631     ERP020507    16S rRNA
#> 3      Catherine Pfister 10.7717/peerj.631     ERP020507    16S rRNA
#> 4      Catherine Pfister 10.7717/peerj.631     ERP020507    16S rRNA
#> 5      Catherine Pfister 10.7717/peerj.631     ERP020507    16S rRNA
#> 6      Catherine Pfister 10.7717/peerj.631     ERP020507    16S rRNA
#>   target_subfragment                                       pcr_primers
#> 1                 V4 FWD:GTGCCAGCMGCCGCGGTAA; REV:GGACTACHVGGGTWTCTAAT
#> 2                 V4 FWD:GTGCCAGCMGCCGCGGTAA; REV:GGACTACHVGGGTWTCTAAT
#> 3                 V4 FWD:GTGCCAGCMGCCGCGGTAA; REV:GGACTACHVGGGTWTCTAAT
#> 4                 V4 FWD:GTGCCAGCMGCCGCGGTAA; REV:GGACTACHVGGGTWTCTAAT
#> 5                 V4 FWD:GTGCCAGCMGCCGCGGTAA; REV:GGACTACHVGGGTWTCTAAT
#> 6                 V4 FWD:GTGCCAGCMGCCGCGGTAA; REV:GGACTACHVGGGTWTCTAAT
#>   illumina_technology extraction_center run_center run_date read_length_bp
#> 1               HiSeq       ANL-Chicago        ANL   6/2/11            140
#> 2               HiSeq       ANL-Chicago        ANL   6/2/11            140
#> 3               HiSeq       ANL-Chicago        ANL   6/2/11            140
#> 4               HiSeq       ANL-Chicago        ANL   6/2/11            140
#> 5               HiSeq       ANL-Chicago        ANL   6/2/11            140
#> 6               HiSeq       ANL-Chicago        ANL   6/2/11            140
#>   sequences_split_libraries observations_closed_ref_greengenes
#> 1                    298351                             275292
#> 2                    281588                             219846
#> 3                    187735                             165334
#> 4                    271744                             221436
#> 5                    230631                             140483
#> 6                    217961                             134035
#>   observations_closed_ref_silva observations_open_ref_greengenes
#> 1                        280849                           297717
#> 2                        230554                           280463
#> 3                        168100                           186813
#> 4                        231443                           270818
#> 5                        167126                           229193
#> 6                        169479                           217037
#>   observations_deblur_90bp observations_deblur_100bp observations_deblur_150bp
#> 1                   201116                    194458                     96458
#> 2                   177503                    174092                     90406
#> 3                   119184                    118020                     64784
#> 4                   181115                    170197                     86880
#> 5                   142987                    138395                     63409
#> 6                   139000                    133015                     68383
#>   emp_release1 qc_filtered subset_10k subset_5k subset_2k sample_taxid
#> 1         TRUE        TRUE       TRUE      TRUE      TRUE       718308
#> 2         TRUE        TRUE       TRUE      TRUE      TRUE       718308
#> 3         TRUE        TRUE       TRUE      TRUE      TRUE       718308
#> 4         TRUE        TRUE       TRUE      TRUE      TRUE       718308
#> 5         TRUE        TRUE       TRUE      TRUE      TRUE       718308
#> 6         TRUE        TRUE       TRUE      TRUE      TRUE       718308
#>   sample_scientific_name host_taxid host_common_name_provided host_common_name
#> 1     biofilm metagenome         NA                                           
#> 2     biofilm metagenome         NA                                           
#> 3     biofilm metagenome         NA                                           
#> 4     biofilm metagenome         NA                                           
#> 5     biofilm metagenome         NA                                           
#> 6     biofilm metagenome         NA                                           
#>   host_scientific_name host_superkingdom host_kingdom host_phylum host_class
#> 1                                                                           
#> 2                                                                           
#> 3                                                                           
#> 4                                                                           
#> 5                                                                           
#> 6                                                                           
#>   host_order host_family host_genus host_species collection_timestamp
#> 1                                                 2009-08-06 00:00:00
#> 2                                                 2009-08-06 00:00:00
#> 3                                                 2009-08-06 00:00:00
#> 4                                                 2009-08-06 00:00:00
#> 5                                                 2009-08-06 00:00:00
#> 6                                                 2009-08-06 00:00:00
#>                        country latitude_deg longitude_deg depth_m altitude_m
#> 1 GAZ:United States of America        48.38       -124.73       0          0
#> 2 GAZ:United States of America        48.38       -124.73       0          0
#> 3 GAZ:United States of America        48.38       -124.73       0          0
#> 4 GAZ:United States of America        48.38       -124.73       0          0
#> 5 GAZ:United States of America        48.38       -124.73       0          0
#> 6 GAZ:United States of America        48.38       -124.73       0          0
#>   elevation_m    env_biome     env_feature env_material envo_biome_0
#> 1       24.73 marine biome intertidal zone      biofilm        biome
#> 2       24.73 marine biome intertidal zone      biofilm        biome
#> 3       24.73 marine biome intertidal zone      biofilm        biome
#> 4       24.73 marine biome intertidal zone      biofilm        biome
#> 5       24.73 marine biome intertidal zone      biofilm        biome
#> 6       24.73 marine biome intertidal zone      biofilm        biome
#>    envo_biome_1 envo_biome_2 envo_biome_3 envo_biome_4 envo_biome_5     empo_0
#> 1 aquatic biome marine biome                                        EMP sample
#> 2 aquatic biome marine biome                                        EMP sample
#> 3 aquatic biome marine biome                                        EMP sample
#> 4 aquatic biome marine biome                                        EMP sample
#> 5 aquatic biome marine biome                                        EMP sample
#> 6 aquatic biome marine biome                                        EMP sample
#>        empo_1 empo_2           empo_3 adiv_observed_otus adiv_chao1
#> 1 Free-living Saline Surface (saline)                 95   116.0833
#> 2 Free-living Saline Surface (saline)                187   263.5556
#> 3 Free-living Saline Surface (saline)                190   230.5294
#> 4 Free-living Saline Surface (saline)                181   261.4545
#> 5 Free-living Saline Surface (saline)                195   283.6667
#> 6 Free-living Saline Surface (saline)                181   202.3030
#>   adiv_shannon adiv_faith_pd temperature_deg_c ph salinity_psu oxygen_mg_per_l
#> 1     3.627853      11.74937                NA NA           NA              NA
#> 2     3.898378      22.86418                NA NA           NA              NA
#> 3     4.520886      22.38092                NA NA           NA              NA
#> 4     4.220427      22.08485                NA NA           NA              NA
#> 5     5.141952      25.27403                NA NA           NA              NA
#> 6     4.554962      22.64086                NA NA           NA              NA
#>   phosphate_umol_per_l ammonium_umol_per_l nitrate_umol_per_l
#> 1                   NA                  NA                 NA
#> 2                   NA                  NA                 NA
#> 3                   NA                  NA                 NA
#> 4                   NA                  NA                 NA
#> 5                   NA                  NA                 NA
#> 6                   NA                  NA                 NA
#>   sulfate_umol_per_l
#> 1                 NA
#> 2                 NA
#> 3                 NA
#> 4                 NA
#> 5                 NA
#> 6                 NA