Sample metadata table

This is a meta-data/mapping file containing related information about samples originated from the Earth Microbiome Project.

sample_metadata_sub2k

Format

Data frame with columns

#SampleID

unique sample identifier (leading hashtag is required by Qiime)

BarcodeSequence

barcode sequence

LinkerPrimerSequence

linker primer sequence

Description

sample description

host_subject_id

unique host or subject (can have multiple samples per host_subject_id)

study_id

study identifier

title

study title

principal_investigator

PI of the study

doi

Digital Object Identifier (DOI) or PubMed ID of primary publication

ebi_accession

European Bioinformatics Institute (EBI) European Nucleotide Archive (ENA) accession number if submitted

target_gene

name of gene amplified (e.g., 16S rRNA)

target_subfragment

name of subfragment of gene amplified (e.g., V4)

pcr_primers

amplicon primer sequences used

illumina_technology

model of Illumina sequencer

extraction_center

where the sample was physically extracted

run_center

where the sample was physically sequenced (CCME=U Colorado Boulder, ANL=Argonne Natl Lab, UCSDMI=UC San Diego, CGS-GL=Wash U, UCD=U Colorado Denver)

run_date

date the sample was physically sequenced

read_length_bp

median read length in bp across study after quality filtering

sequences_split_libraries

number of sequences after demultiplexing with split_libraries_fastq.py

observations_closed_ref_greengenes

number of observations in closed-reference Greengenes table

observations_closed_ref_silva

number of observations in closed-reference Silva table

observations_open_ref_greengenes

number of observations in open-reference Greengenes table

observations_deblur_90bp

number of observations in 90-bp Deblur table

observations_deblur_100bp

number of observations in 100-bp Deblur table

observations_deblur_150bp

number of observations in 150-bp Deblur table

emp_release1

samples with >=1 sequences (split libraries) per sample

qc_filtered

samples with >=1000 observations in CR-GG & CR-Silva & OR-GG & Deblur-90 but excluding controls (all subsets are in this set)

subset_10k

10000 samples with >=10000 CR-GG & CR-Silva & OR-GG and >=5000 Deblur-90, randomly selected and evenly distributed across empo_3 categories and then across studies

subset_5k

5000 samples with >=10000 CR-GG & CR-Silva & OR-GG and >=5000 Deblur-90, randomly selected and evenly distributed across empo_3 categories and then across studies

subset_2k

2000 samples with >=10000 CR-GG & CR-Silva & OR-GG and >=5000 Deblur-90, randomly selected and evenly distributed across empo_3 categories and then across studies

sample_taxid

sample NCBI taxonomy ID

sample_scientific_name

sample NCBI scientific name looked up from taxonomy ID

host_taxid

host NCBI taxonomy ID

host_common_name_provided

host common name provided in the original mapping file

host_common_name

host NCBI common name looked up from taxonomy ID

host_scientific_name

host NCBI scientific name looked up from taxonomy ID

host_superkingdom

host superkingdom looked up from taxonomy ID

host_kingdom

host kingdom looked up from taxonomy ID

host_phylum

host phylum looked up from taxonomy ID

host_class

host class looked up from taxonomy ID

host_order

host order looked up from taxonomy ID

host_family

host family looked up from taxonomy ID

host_genus

host genus looked up from taxonomy ID

host_species

host species looked up from taxonomy ID

collection_timestamp

date and time when sample was collected

country

country where sample was collected

latitude_deg

latitude in degrees

longitude_deg

longitude in degrees

depth_m

depth in meters of sample below surface (earth surface if soil, sea/lake bottom if sediment, lake surface if lake, sea level if marine)

altitude_m

height above surface, usually zero unless the mouse is levitating (either depth or altitude can have a non-zero value but not both)

elevation_m

height above sea level in meters (from georeferencing tool)

env_biome

ENVO biome

env_feature

ENVO feature

env_material

ENVO material

envo_biome_0

level 0 ENVO biome looked up from given ENVO biome

envo_biome_1

level 1 ENVO biome looked up from given ENVO biome

envo_biome_2

level 2 ENVO biome looked up from given ENVO biome

envo_biome_3

level 3 ENVO biome looked up from given ENVO biome

envo_biome_4

level 4 ENVO biome looked up from given ENVO biome

envo_biome_5

level 5 ENVO biome looked up from given ENVO biome

empo_0

level 0 EMPO habitat from EMPO ontology

empo_1

level 1 EMPO habitat from EMPO ontology

empo_2

level 2 EMPO habitat from EMPO ontology

empo_3

level 3 EMPO habitat from EMPO ontology

adiv_observed_otus

observed tag sequences of 90-bp Deblur table rarefied to 5000 sequences per sample

adiv_chao1

Chao1 index of 90-bp Deblur table rarefied to 5000 sequences per sample

adiv_shannon

Shannon index of 90-bp Deblur table rarefied to 5000 sequences per sample

adiv_faith_pd

Faith's phylogenetic diversity of 90-bp Deblur table rarefied to 5000 sequences per sample

temperature_deg_c

ambient temperature in degrees Celcius

ph

pH value

salinity_psu

salinity in practical salinity units (PSU)

oxygen_mg_per_l

oxygen concentration in mg/L

phosphate_umol_per_l

phosphate concentration in µmol/L

ammonium_umol_per_l

ammonium concentration in µmol/L

nitrate_umol_per_l

nitrate concentration in µmol/L

sulfate_umol_per_l

sulfate concentration in µmol/L

Source

https://earthmicrobiome.org/

Examples

  str(sample_metadata_sub2k)

#> Classes 'data.table' and 'data.frame':	2000 obs. of  76 variables:
#>  $ #SampleID                         : chr  "550.L1S116.s.1.sequence" "550.L1S119.s.1.sequence" "550.L1S164.s.1.sequence" "550.L1S194.s.1.sequence" ...
#>  $ BarcodeSequence                   : chr  "ATGCCTGAGCAG" "CAGCACTAAGCG" "ATGTACGGCGAC" "CGAAGACTGCTG" ...
#>  $ LinkerPrimerSequence              : chr  "GTGCCAGCMGCCGCGGTAA" "GTGCCAGCMGCCGCGGTAA" "GTGCCAGCMGCCGCGGTAA" "GTGCCAGCMGCCGCGGTAA" ...
#>  $ Description                       : chr  "sample_20 stool" "sample_23 stool" "sample_73 stool" "sample_105 stool" ...
#>  $ host_subject_id                   : chr  "F4" "F4" "M3" "M3" ...
#>  $ study_id                          : int  550 550 550 550 550 550 550 550 550 550 ...
#>  $ title                             : chr  "Moving pictures of the human microbiome" "Moving pictures of the human microbiome" "Moving pictures of the human microbiome" "Moving pictures of the human microbiome" ...
#>  $ principal_investigator            : chr  "Rob Knight" "Rob Knight" "Rob Knight" "Rob Knight" ...
#>  $ doi                               : chr  "10.1186/gb-2011-12-5-r50" "10.1186/gb-2011-12-5-r50" "10.1186/gb-2011-12-5-r50" "10.1186/gb-2011-12-5-r50" ...
#>  $ ebi_accession                     : chr  "ERP021896" "ERP021896" "ERP021896" "ERP021896" ...
#>  $ target_gene                       : chr  "16S rRNA" "16S rRNA" "16S rRNA" "16S rRNA" ...
#>  $ target_subfragment                : chr  "V4" "V4" "V4" "V4" ...
#>  $ pcr_primers                       : chr  "FWD:GTGCCAGCMGCCGCGGTAA; REV:GGACTACHVGGGTWTCTAAT" "FWD:GTGCCAGCMGCCGCGGTAA; REV:GGACTACHVGGGTWTCTAAT" "FWD:GTGCCAGCMGCCGCGGTAA; REV:GGACTACHVGGGTWTCTAAT" "FWD:GTGCCAGCMGCCGCGGTAA; REV:GGACTACHVGGGTWTCTAAT" ...
#>  $ illumina_technology               : chr  "HiSeq" "HiSeq" "HiSeq" "HiSeq" ...
#>  $ extraction_center                 : chr  "CCME-Boulder" "CCME-Boulder" "CCME-Boulder" "CCME-Boulder" ...
#>  $ run_center                        : chr  "CGS-GL" "CGS-GL" "CGS-GL" "CGS-GL" ...
#>  $ run_date                          : chr  "8/27/10" "8/27/10" "8/27/10" "8/27/10" ...
#>  $ read_length_bp                    : int  132 132 132 132 132 132 132 132 132 132 ...
#>  $ sequences_split_libraries         : int  33383 40944 35636 46992 30131 29907 40565 49669 16954 20378 ...
#>  $ observations_closed_ref_greengenes: int  32153 39472 34550 43925 29179 29034 39123 47800 15824 18484 ...
#>  $ observations_closed_ref_silva     : int  32453 39929 34666 43852 29553 29310 39200 48154 16107 19185 ...
#>  $ observations_open_ref_greengenes  : int  33337 40870 35599 46875 30094 29865 40433 49582 16870 20280 ...
#>  $ observations_deblur_90bp          : int  22567 27871 24134 30041 21132 20875 25919 32962 10838 13559 ...
#>  $ observations_deblur_100bp         : int  22160 27191 23686 29264 20643 20401 25653 32301 10983 13608 ...
#>  $ observations_deblur_150bp         : int  1043 1272 1161 1974 603 572 4033 1602 391 913 ...
#>  $ emp_release1                      : logi  TRUE TRUE TRUE TRUE TRUE TRUE ...
#>  $ qc_filtered                       : logi  TRUE TRUE TRUE TRUE TRUE TRUE ...
#>  $ subset_10k                        : logi  TRUE TRUE TRUE TRUE TRUE TRUE ...
#>  $ subset_5k                         : logi  TRUE TRUE TRUE TRUE TRUE TRUE ...
#>  $ subset_2k                         : logi  TRUE TRUE TRUE TRUE TRUE TRUE ...
#>  $ sample_taxid                      : int  408170 408170 408170 408170 408170 408170 408170 408170 539655 539655 ...
#>  $ sample_scientific_name            : chr  "human gut metagenome" "human gut metagenome" "human gut metagenome" "human gut metagenome" ...
#>  $ host_taxid                        : int  9606 9606 9606 9606 9606 9606 9606 9606 9606 9606 ...
#>  $ host_common_name_provided         : chr  "human" "human" "human" "human" ...
#>  $ host_common_name                  : chr  "human" "human" "human" "human" ...
#>  $ host_scientific_name              : chr  "Homo sapiens" "Homo sapiens" "Homo sapiens" "Homo sapiens" ...
#>  $ host_superkingdom                 : chr  "sk__Eukaryota" "sk__Eukaryota" "sk__Eukaryota" "sk__Eukaryota" ...
#>  $ host_kingdom                      : chr  "k__Metazoa" "k__Metazoa" "k__Metazoa" "k__Metazoa" ...
#>  $ host_phylum                       : chr  "p__Chordata" "p__Chordata" "p__Chordata" "p__Chordata" ...
#>  $ host_class                        : chr  "c__Mammalia" "c__Mammalia" "c__Mammalia" "c__Mammalia" ...
#>  $ host_order                        : chr  "o__Primates" "o__Primates" "o__Primates" "o__Primates" ...
#>  $ host_family                       : chr  "f__Hominidae" "f__Hominidae" "f__Hominidae" "f__Hominidae" ...
#>  $ host_genus                        : chr  "g__Homo" "g__Homo" "g__Homo" "g__Homo" ...
#>  $ host_species                      : chr  "s__Homo_sapiens" "s__Homo_sapiens" "s__Homo_sapiens" "s__Homo_sapiens" ...
#>  $ collection_timestamp              : POSIXct, format: "2009-03-29" "2009-04-06" ...
#>  $ country                           : chr  "GAZ:United States of America" "GAZ:United States of America" "GAZ:United States of America" "GAZ:United States of America" ...
#>  $ latitude_deg                      : num  40 40 40 40 40 ...
#>  $ longitude_deg                     : num  -105 -105 -105 -105 -105 ...
#>  $ depth_m                           : num  0 0 0 0 0 0 0 0 0 0 ...
#>  $ altitude_m                        : num  0 0 0 0 0 0 0 0 0 0 ...
#>  $ elevation_m                       : num  1624 1624 1624 1624 1624 ...
#>  $ env_biome                         : chr  "urban biome" "urban biome" "urban biome" "urban biome" ...
#>  $ env_feature                       : chr  "human-associated habitat" "human-associated habitat" "human-associated habitat" "human-associated habitat" ...
#>  $ env_material                      : chr  "feces" "feces" "feces" "feces" ...
#>  $ envo_biome_0                      : chr  "biome" "biome" "biome" "biome" ...
#>  $ envo_biome_1                      : chr  "terrestrial biome" "terrestrial biome" "terrestrial biome" "terrestrial biome" ...
#>  $ envo_biome_2                      : chr  "anthropogenic terrestrial biome" "anthropogenic terrestrial biome" "anthropogenic terrestrial biome" "anthropogenic terrestrial biome" ...
#>  $ envo_biome_3                      : chr  "dense settlement biome" "dense settlement biome" "dense settlement biome" "dense settlement biome" ...
#>  $ envo_biome_4                      : chr  "urban biome" "urban biome" "urban biome" "urban biome" ...
#>  $ envo_biome_5                      : chr  "" "" "" "" ...
#>  $ empo_0                            : chr  "EMP sample" "EMP sample" "EMP sample" "EMP sample" ...
#>  $ empo_1                            : chr  "Host-associated" "Host-associated" "Host-associated" "Host-associated" ...
#>  $ empo_2                            : chr  "Animal" "Animal" "Animal" "Animal" ...
#>  $ empo_3                            : chr  "Animal distal gut" "Animal distal gut" "Animal distal gut" "Animal distal gut" ...
#>  $ adiv_observed_otus                : int  93 82 108 107 79 76 173 125 627 396 ...
#>  $ adiv_chao1                        : num  114.4 116.5 135.1 126 94.5 ...
#>  $ adiv_shannon                      : num  3.87 3.27 3.66 4.44 3.33 ...
#>  $ adiv_faith_pd                     : num  12.5 10.7 14.2 12 11.8 ...
#>  $ temperature_deg_c                 : num  NA NA NA NA NA NA NA NA NA NA ...
#>  $ ph                                : num  NA NA NA NA NA NA NA NA NA NA ...
#>  $ salinity_psu                      : num  NA NA NA NA NA NA NA NA NA NA ...
#>  $ oxygen_mg_per_l                   : num  NA NA NA NA NA NA NA NA NA NA ...
#>  $ phosphate_umol_per_l              : num  NA NA NA NA NA NA NA NA NA NA ...
#>  $ ammonium_umol_per_l               : num  NA NA NA NA NA NA NA NA NA NA ...
#>  $ nitrate_umol_per_l                : num  NA NA NA NA NA NA NA NA NA NA ...
#>  $ sulfate_umol_per_l                : num  NA NA NA NA NA NA NA NA NA NA ...
#>  - attr(*, ".internal.selfref")=<externalptr>